+++ renderer/components/index.ts
@@ -46,3 +46,4 @@ export { SeedingDashboard } from './SeedingDashboard';
 export { PrivacySettings } from './PrivacySettings';
 export { TorrentControlModal } from './TorrentControlModal';
 export { StreamPlayerModal } from './StreamPlayerModal';
+export { ShareLinkModal } from './ShareLinkModal';